Added RfCompression and support for cmpType 1
Compression has been refactored and moved to a static class `LSRutil.RF.RfCompression`. This class handles wrapping non-0 compression methods around the original file size field and then handling the underlying compression method.

Compression type 1 is Huffman coding. Where symbols are separated by individual byte values. The bitstream for Huffman coding is written using little endian 32-bit integer blocks. The Huffman coding tree is stored using an initial 16-bit integer codeword, followed by a 1024-value tree of 16-bit codewords. (The first 256 values are unused, and refer to original byte values).

* Compression is now supported for `ResourceFile.Pack` operations.
* Added overloads to `ResourceFile.Pack` and `ResourceArchive.Add` that take `ResourceCompressionType` instead of `bool compress=true`.
* The underlying type of the `ResourceFile.compressionType` field has been changed to `ResourceCompressionType`.
* An `InvalidDataException` is now thrown during `ResourceFile.Pack`/`ResourceFile.Unpack` for files sporting an invalid compression type (not 0, 1, or 2).
* Added commented-out `ResourceCompressionType` parameters for the WriteRF example program.

+++ RF/RfWriter.cs
@@ -32,7 +32,7 @@ namespace LSRutil.RF
                 {
                     headerWriter.Write(resFile.filepath.Length+1); // Compensates for null terminator to be added.
                     headerWriter.Write((uint)((DateTimeOffset)resFile.timestamp).ToUnixTimeSeconds());
-                    headerWriter.Write(resFile.compressionType);
+                    headerWriter.Write((int)resFile.compressionType);
                     headerWriter.Write(resFile.compressedSize);
                     headerWriter.Write(resFile.offset);
                     headerWriter.Write(Encoding.ASCII.GetBytes(resFile.filepath+"\0"));
